+++ CHANGELOG.md
@@ -184,6 +184,7 @@
 * [X] breeze-logger 完成自己的日志器的实现。
 * [X] breeze-core 添加`String.alignLeft`等方法。
 * [X] breeze-core 添加`String.takeIfNotEmpty`和`String.takeIfNotBlank`。 
+* [X] breeze-core 去除以上扩展，添加`String.orNull`和`Optional.orNull`扩展。
 ***
 * [ ] breeze-dsl 优化：对于表示转换的元素，可以通过`"a"(...) fromTo "b"`的语法构建。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）